movie_genre_classifier_project/requirements.txt
streamlit==1.28.0
pandas==2.0.3
requests==2.31.0
plotly==5.15.0
python-dotenv==1.0.0
openpyxl==3.1.2
scikit-learn==1.3.2